Refactor API routes for graph configuration: remove unused POST method, add new config route, and update query handling in GET method. Update URLs in e2e tests and improve string interpolation in API calls.
