+++ app/api/graph/[graph]/route.ts
@@ -2,7 +2,6 @@ import { NextRequest, NextResponse } from "next/server";
 import { getClient } from "@/app/api/auth/[...nextauth]/options";
 import { prepareArg, securedFetch } from "@/lib/utils";
 
-// eslint-disable-next-line import/prefer-default-export
 export async function DELETE(request: NextRequest, { params }: { params: Promise<{ graph: string }> }) {
 
     const session = await getClient()
@@ -29,7 +28,6 @@ export async function DELETE(request: NextRequest, { params }: { params: Promise
     }
 }
 
-// eslint-disable-next-line import/prefer-default-export
 export async function POST(request: NextRequest, { params }: { params: Promise<{ graph: string }> }) {
 
     const session = await getClient()
@@ -85,7 +83,6 @@ export async function POST(request: NextRequest, { params }: { params: Promise<{
     }
 }
 
-// eslint-disable-next-line import/prefer-default-export
 export async function PATCH(request: NextRequest, { params }: { params: Promise<{ graph: string }> }) {
 
     const session = await getClient()
@@ -122,22 +119,11 @@ export async function GET(request: NextRequest, { params }: { params: Promise<{ 
     const { client, user } = session
 
     const { graph: graphId } = await params
-
-    const ID = request.nextUrl.searchParams.get("ID")
+    const query = request.nextUrl.searchParams.get("query")
+    const create = request.nextUrl.searchParams.get("create")
+    const { role } = user
 
     try {
-        if (ID) {
-            // const result = await securedFetch(`https://localhost:5000/progress/?ID=${ID}`, {
-            //     method: "GET"
-            // })
-            // if (!result.ok) throw new Error("something went wrong")
-            // const json = await result.json()
-            return NextResponse.json({ progress: 10 }, { status: 200 })
-        }
-
-        const query = request.nextUrl.searchParams.get("query")
-        const create = request.nextUrl.searchParams.get("create")
-        const { role } = user
 
         if (!query) throw new Error("Missing parameter query")
